Add monaco editor autocomplete for set statement
